Resnet/requirements.txt
numpy==1.24.1
Pillow==9.3.0
tqdm==4.65.0
PrettyTable=3.7.0
